dynamic master - Cluster and Gateway Master handling

+++ nexus_wipe_messages.sh
@@ -7,14 +7,20 @@ source colors_set.sh
 # If no template name to be used to create .env use 'default'
 TEMPLATE=${1-default}
 
-# Shut down actual container composition
-echo -e "${PURPLE}Shut down actual container composition${NC}"
-docker compose down
+# Bounce docker (here down) if option is given
+if [ $1 == "-d" ] ; then
+  # Shut down actual container composition
+  echo -e "${PURPLE}Shut down actual container composition${NC}"
+  docker compose down
+fi
 
 # Remove Nexus Message Store file
 echo -e "${PURPLE}Remove ./nexus_root/.nexus/storage/messages.umspack${NC}"
 sudo rm -f ./nexus_root/.nexus/storage/messages.umsgpack
 
-# Launch container composition
-echo -e "${PURPLE}Launch container composition${NC}"
-docker compose up -d
+# Bounce docker (here up) if option is given
+if [ $1 == "-d" ] ; then
+  # Launch container composition
+  echo -e "${PURPLE}Launch container composition${NC}"
+  docker compose up -d
+fi
